chore: trim chat error diagnostics diff
