+++ site/src/pages/AgentsPage/components/ChatConversation/LiveStreamTail.stories.tsx
@@ -328,9 +328,7 @@ export const GenericErrorDoesNotShowUsageAction: Story = {
 		liveStatus: buildLiveStatus({
 			persistedError: {
 				kind: "generic",
-				message: "The chat request failed unexpectedly.",
-				detail:
-					'stream response: Post "https://api.anthropic.com/v1/messages": decoder failed',
+				message: "Provider request failed.",
 			},
 		}),
 	},
@@ -339,10 +337,7 @@ export const GenericErrorDoesNotShowUsageAction: Story = {
 		expect(
 			canvas.getByRole("heading", { name: /request failed/i }),
 		).toBeVisible();
-		expect(
-			canvas.getByText(/the chat request failed unexpectedly/i),
-		).toBeVisible();
-		expect(canvas.getByText(/decoder failed/i)).toBeVisible();
+		expect(canvas.getByText(/provider request failed/i)).toBeVisible();
 		expect(
 			canvas.queryByText(/start a conversation with your agent/i),
 		).not.toBeInTheDocument();
